Added Name, Removed and added socials, added desc
I still have more work to do on this. I need to add more to the front description about my self and get that changed. This will take some time to do thou.  Also, we need to have it point to the images that I have. Also we need to get someone to get those images changed as well.  Just send it to someone and they can change it or add some variety to it.  Also, we may need to change the color theme as well.