Added My Projects to the Projects List
Aslo made some Tweaks to the about page and the toolstack page. I added some tools to the list and OS to it as well. There are a few more Projects I need to add to it then I am done.

I am not sure, but I might change the color scheme of the website to blue and white, then take it from there.